+++ src/Pages/MyPage/Components/MyPageMain.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from "react";
 import styled from "styled-components";
-import Product from "../../../../Components/Product";
-import Title from "../Title/Title";
-import Blank from "../Blank/Blank";
+import Product from "../../../Components/Product";
+import Title from "./Title";
+import Blank from "./Blank";
 import { useHistory } from "react-router";
 
 function MyPageMain({ productLike, productBuy, productCreate, queryString }) {
