--- src/Pages/MyPage/Pages/MyPageBuy/MyPageBuy.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function MyPageBuy(props) {
-  return <div></div>;
-}
-
-export default MyPageBuy;
