--- src/Pages/MyPage/Pages/MyPageCreate/MyPageCreate.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function MyPageCreate(props) {
-  return <div></div>;
-}
-
-export default MyPageCreate;
