--- src/Pages/MyPage/Pages/MyPageLike/MyPageLike.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function MyPageLike(props) {
-  return <div></div>;
-}
-
-export default MyPageLike;
